make preview loaders less dependent
